+++ conkit/io/a3m.py
@@ -72,7 +72,7 @@ class A3mParser(SequenceFileParser):
 
         Returns
         -------
-        :obj:`SequenceFile <conkit.core.sequencefile.SequenceFile>`
+        :obj:`~conkit.core.sequencefile.SequenceFile`
 
         """
 
@@ -164,7 +164,7 @@ class A3mParser(SequenceFileParser):
         ----------
         f_handle
            Open file handle [write permissions]
-        hierarchy : :obj:`SequenceFile <conkit.core.sequencefile.SequenceFile>`, :obj:`Sequence <conkit.core.sequence.Sequence>`
+        hierarchy : :obj:`~conkit.core.sequencefile.SequenceFile`, :obj:`~conkit.core.sequence.Sequence`
 
         """
         # Double check the type of hierarchy and reconstruct if necessary