+++ conkit/io/evfold.py
@@ -49,6 +49,7 @@ RE_SPLIT = re.compile(r'\s+')
 class EVfoldParser(ContactFileParser):
     """Class to parse a EVfold contact file
     """
+
     def __init__(self):
         super(EVfoldParser, self).__init__()
 
@@ -64,7 +65,7 @@ class EVfoldParser(ContactFileParser):
 
         Returns
         -------
-        :obj:`ContactFile <conkit.core.contactfile.ContactFile>`
+        :obj:`~conkit.core.contactfile.ContactFile`
 
         """
 
@@ -81,11 +82,7 @@ class EVfoldParser(ContactFileParser):
             else:
                 res1_seq, res1, res2_seq, res2, _, raw_score = RE_SPLIT.split(line)
 
-                contact = Contact(
-                    int(res1_seq),
-                    int(res2_seq),
-                    float(raw_score)
-                )
+                contact = Contact(int(res1_seq), int(res2_seq), float(raw_score))
                 contact.res1 = res1
                 contact.res2 = res2
                 contact_map.add(contact)
@@ -101,12 +98,12 @@ class EVfoldParser(ContactFileParser):
         ----------
         f_handle
            Open file handle [write permissions]
-        hierarchy : :obj:`ContactFile <conkit.core.contactfile.ContactFile>`, :obj:`ContactMap <conkit.core.contactmap.ContactMap>`
-                    or :obj:`Contact <conkit.core.contact.Contact>`
+        hierarchy : :obj:`~conkit.core.contactfile.ContactFile`, :obj:`~conkit.core.contactmap.ContactMap`
+                    or :obj:`~conkit.core.contact.Contact`
 
         Raises
         ------
-        RuntimeError
+        :obj:`RuntimeError`
            More than one contact map in the hierarchy
 
         """
@@ -121,8 +118,12 @@ class EVfoldParser(ContactFileParser):
         for contact_map in contact_file:
             for contact in contact_map:
                 line = "{res1_seq} {res1} {res2_seq} {res2} 0 {raw_score}"
-                line = line.format(res1_seq=contact.res1_seq, res2_seq=contact.res2_seq,
-                                   res1=contact.res1, res2=contact.res2, raw_score=contact.raw_score)
+                line = line.format(
+                    res1_seq=contact.res1_seq,
+                    res2_seq=contact.res2_seq,
+                    res1=contact.res1,
+                    res2=contact.res2,
+                    raw_score=contact.raw_score)
                 content += line + os.linesep
 
         f_handle.write(content)