+++ conkit/plot/sequencecoverage.py
@@ -58,7 +58,7 @@ class SequenceCoverageFigure(Figure):
 
     Attributes
     ----------
-    hierarchy : :obj:`SequenceFile <conkit.core.sequencefile.SequenceFile>`
+    hierarchy : :obj:`~conkit.core.sequencefile.SequenceFile`
        The Multiple Sequence Alignment hierarchy
 
     Examples
@@ -74,10 +74,10 @@ class SequenceCoverageFigure(Figure):
 
         Parameters
         ----------
-        hierarchy : :obj:`SequenceFile <conkit.core.sequencefile.SequenceFile>`
+        hierarchy : :obj:`~conkit.core.sequencefile.SequenceFile`
            The Multiple Sequence Alignment hierarchy
         **kwargs
-           General :obj:`Figure <conkit.plot.figure.Figure>` keyword arguments
+           General :obj:`~conkit.plot.figure.Figure` keyword arguments
 
         """
         super(SequenceCoverageFigure, self).__init__(**kwargs)
@@ -92,12 +92,12 @@ class SequenceCoverageFigure(Figure):
 
     @property
     def hierarchy(self):
-        """A ConKit :obj:`SequenceFile <conkit.core.sequencefile.SequenceFile>`"""
+        """A ConKit :obj:`~conkit.core.sequencefile.SequenceFile`"""
         return self._hierarchy
 
     @hierarchy.setter
     def hierarchy(self, hierarchy):
-        """Define the ConKit :obj:`SequenceFile <conkit.core.sequencefile.SequenceFile>` """
+        """Define the ConKit :obj:`~conkit.core.sequencefile.SequenceFile` """
         if hierarchy and _isinstance(hierarchy, "SequenceFile") and hierarchy.is_alignment:
             self._hierarchy = hierarchy
         elif hierarchy and _isinstance(hierarchy, "SequenceFile"):